+++ extensions/vscode/src/multiStepInputs/newDeployment.ts
@@ -943,6 +943,7 @@ export async function newDeployment(
         newDeploymentData.newCredentials.url,
         newDeploymentData.newCredentials.apiKey,
         newDeploymentData.newCredentials.snowflakeConnection,
+        serverType,
       );
       newOrSelectedCredential = response.data;
     } catch (error: unknown) {