+++ internal/clients/connect_cloud/user.go
@@ -1,0 +1,6 @@
+package connect_cloud
+
+// Copyright (C) 2025 by Posit Software, PBC.
+
+type UserResponse struct {
+}